pyproject.toml
[tool.poetry]
name = "apollo11-simulator"
version = "1.0.0"
description = ""
authors = ["Erick Calderin", "Anabel Arteaga", "Pablo Alarcon"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.5.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
